requirements.txt
fastapi
uvicorn[standard]
langchain
chromadb
sentence-transformers
PyMuPDF
python-docx
jinja2
pydantic
python-multipart 
openai 
python-dotenv
email-validator
passlib[bcrypt]
python-jose[cryptography]
PyJWT
pymongo
motor